implementing PR

+++ src/main/scala/com/github/tartakynov/osm/algorithms/UpstreamSegmentsCounter.scala
@@ -15,10 +15,10 @@ object UpstreamSegmentsCounter extends WeightsCalculator with StrictLogging {
     logger.info(s"Calculating segments weights. Found ${graph.size} segments")
     val startTime = System.currentTimeMillis()
     val segments = graph.keySet.toList.sortBy(segment => graph(segment).count(_.flowsInto(segment)))
-    val weights = mutable.Map[Segment, Int]()
+    val weights = mutable.Map[Segment, Double]()
     val visited = mutable.Set[Segment]()
     for (segment <- segments) {
-      val stack = mutable.Stack[(Segment, Int)]()
+      val stack = mutable.Stack[(Segment, Double)]()
       if (!visited.contains(segment)) {
         stack.push((segment, 1))
       }
@@ -27,7 +27,7 @@ object UpstreamSegmentsCounter extends WeightsCalculator with StrictLogging {
         val (current, w) = stack.pop()
         if (!visited.contains(current)) {
           visited.add(current)
-          weights.update(current, weights.getOrElse(current, 0) + w)
+          weights.update(current, weights.getOrElse(current, 0d) + w)
           val links = graph(current).filter(current.flowsInto)
           stack.pushAll(links.map((_, w + 1)))
         }
