+++ src/main/scala/com/github/tartakynov/osm/algorithms/WeightsCalculator.scala
@@ -4,7 +4,7 @@ import com.github.tartakynov.osm.graph.Graph.Graph
 import com.github.tartakynov.osm.graph.Segment
 
 trait WeightsCalculator {
-  type Weights = Map[Segment, Int]
+  type Weights = Map[Segment, Double]
 
   def calculate(graph: Graph): Weights
 }
